# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /account:
    get:
      operationId: getAccount
      tags: [account]
      summary: The caller's profile, preferences, and subscription summary
      description: >-
        Returns the authenticated learner's profile (name, email, avatar), their
        notification and learning preferences, and a compact subscription
        summary — everything the account settings screen renders from.
      responses:
        "200":
          description: The caller's account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteAccount
      tags: [account]
      summary: Permanently delete the caller's account and all their data
      description: >-
        Permanently deletes the learner's Clerk identity and every local record
        (attempts, progress, friendships, subscription state). Idempotent and
        irreversible; the client signs the user out afterwards.
      responses:
        "200":
          description: The account was deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAccountResult"
  /account/profile:
    patch:
      operationId: updateAccountProfile
      tags: [account]
      summary: Update the caller's display name and/or avatar
      description: >-
        Updates the display name (mirrored to Clerk, the identity source of
        truth) and/or the avatar reference. Only the fields present in the body
        are changed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: The updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountProfileResult"
        "400":
          description: Nothing to update or an invalid field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: The identity service could not apply the change
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /account/preferences:
    patch:
      operationId: updateAccountPreferences
      tags: [account]
      summary: Update notification and/or learning preferences
      description: >-
        Updates any subset of the notification (daily reminder toggle + time)
        and learning (active language, daily goal, theme) preferences. Only the
        provided fields change; each is validated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePreferencesInput"
      responses:
        "200":
          description: The updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountPreferencesResult"
        "400":
          description: An invalid preference value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: The chosen language does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /account/subscription:
    get:
      operationId: getAccountSubscription
      tags: [account]
      summary: The caller's full subscription-management snapshot
      description: >-
        Returns the tier/status, relevant dates, chosen language, a best-effort
        payment-method summary, and billing/invoice history. Softer fields
        (payment method, history) are pulled from the provider where available
        and degrade to null/empty when it isn't configured or doesn't expose
        them.
      responses:
        "200":
          description: The subscription details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionDetails"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /account/subscription/cancel:
    post:
      operationId: cancelAccountSubscription
      tags: [account]
      summary: Cancel the caller's subscription at the period boundary
      description: >-
        Marks the subscription canceled. Paid access continues until the current
        period ends (the plan stays live until currentPeriodEnd lapses).
        Canceling also clears any active pause.
      responses:
        "200":
          description: The updated subscription details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionDetails"
        "400":
          description: No active subscription to cancel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /account/subscription/pause:
    post:
      operationId: pauseAccountSubscription
      tags: [account]
      summary: Pause the caller's subscription for a bounded window
      description: >-
        Pauses the subscription for 1–3 months. While paused the provider
        suspends paid access but does not expire the subscription; it resumes
        automatically when the window closes.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PauseSubscriptionInput"
      responses:
        "200":
          description: The updated subscription details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionDetails"
        "400":
          description: No active subscription to pause or an invalid month count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The subscription is already paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /account/subscription/resume:
    post:
      operationId: resumeAccountSubscription
      tags: [account]
      summary: Resume a canceling subscription (undo a pending cancel)
      description: >-
        Clears a pending cancellation, returning the subscription status from
        canceled to active. No discount is applied, the retention offer is not
        consumed, and the current period end is unchanged. Idempotent for an
        already-active paid subscription and repeatable any number of times.
      responses:
        "200":
          description: The updated subscription details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionDetails"
        "400":
          description: No paid subscription to resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The subscription is paused, not canceling
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /account/subscription/unpause:
    post:
      operationId: unpauseAccountSubscription
      tags: [account]
      summary: Resume a paused subscription early
      description: >-
        Lets a learner who changes their mind come back before the pause
        window closes. Clears the pause and resumes the underlying paid tier
        immediately — the same outcome as waiting for pauseUntil to elapse
        naturally, just early. Only applies to a currently-paused
        subscription; a canceling, active, or expired subscription has
        nothing to unpause.
      responses:
        "200":
          description: The updated subscription details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionDetails"
        "400":
          description: The subscription is not paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /account/subscription/retention:
    post:
      operationId: acceptRetentionOffer
      tags: [account]
      summary: Accept the one-time discounted 3-month retention offer
      description: >-
        Redeems the discounted 3-month retention offer. Resumes/keeps the paid
        tier (clearing any pending cancel or pause), extends the current period
        by three months, and records that the offer was accepted so it can only
        be redeemed once.
      responses:
        "200":
          description: The updated subscription details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionDetails"
        "400":
          description: No subscription eligible for the retention offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The retention offer was already redeemed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AccountProfile:
      type: object
      description: The learner's identity fields, mirrored from Clerk.
      properties:
        id:
          type: string
        email:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
      required:
        - id
        - email
        - displayName
        - avatarUrl
...
    AccountPreferences:
      type: object
      properties:
        notifications:
          $ref: "#/components/schemas/AccountNotificationPreferences"
        learning:
          $ref: "#/components/schemas/AccountLearningPreferences"
      required:
        - notifications
        - learning
...
    PauseSubscriptionInput:
      type: object
      description: How long to pause the subscription for.
      properties:
        months:
          type: integer
          minimum: 1
          maximum: 3
          description: Number of months to pause (1–3). Defaults to 1 if omitted.
    Account:
      type: object
      properties:
        profile:
          $ref: "#/components/schemas/AccountProfile"
        preferences:
          $ref: "#/components/schemas/AccountPreferences"
        subscription:
          $ref: "#/components/schemas/AccountSubscriptionSummary"
      required:
        - profile
        - preferences
        - subscription
    AccountProfileResult:
      type: object
      properties:
        profile:
          $ref: "#/components/schemas/AccountProfile"
      required:
        - profile
    AccountPreferencesResult:
      type: object
      properties:
        preferences:
          $ref: "#/components/schemas/AccountPreferences"
      required:
        - preferences
    UpdateProfileInput:
      type: object
      description: Any subset of the editable profile fields.
      properties:
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
    UpdatePreferencesInput:
      type: object
      description: Any subset of the notification and learning preferences.
      properties:
        dailyReminderEnabled:
          type: boolean
        dailyReminderTime:
          type: string
          nullable: true
        activeLanguage:
          type: string
          nullable: true
        dailyGoal:
          type: integer
        theme:
          type: string
          enum: [system, light, dark]
        timezone:
          type: string
          nullable: true
        hasCompletedTour:
          type: boolean
    DeleteAccountResult:
      type: object
      properties:
        deleted:
          type: boolean
      required:
        - deleted
...